+++ src/main/java/com/btl/server/dto/PhieuTinhTienDTO.java
@@ -1,15 +1,23 @@
 package com.btl.server.dto;
 
+import java.math.BigDecimal;
+
 public class PhieuTinhTienDTO {
     private Long phongId;
     private Integer thang;
     private Integer nam;
-    private Double giaPhong;
+    
+    private BigDecimal giaPhong;
+    
     private Integer soDienDung;
-    private Double tienDien;
+    
+    private BigDecimal tienDien;
+    
     private Integer soNuocDung;
-    private Double tienNuoc;
-    private Double tongTien;
+    
+   private BigDecimal tienNuoc;
+    
+    private BigDecimal tongTien;
 
     public Long getPhongId() { return phongId; }
     public void setPhongId(Long phongId) { this.phongId = phongId; }
@@ -20,21 +28,21 @@ public class PhieuTinhTienDTO {
     public Integer getNam() { return nam; }
     public void setNam(Integer nam) { this.nam = nam; }
 
-    public Double getGiaPhong() { return giaPhong; }
-    public void setGiaPhong(Double giaPhong) { this.giaPhong = giaPhong; }
+    public BigDecimal getGiaPhong() { return giaPhong; }
+    public void setGiaPhong(BigDecimal giaPhong) { this.giaPhong = giaPhong; }
 
     public Integer getSoDienDung() { return soDienDung; }
     public void setSoDienDung(Integer soDienDung) { this.soDienDung = soDienDung; }
 
-    public Double getTienDien() { return tienDien; }
-    public void setTienDien(Double tienDien) { this.tienDien = tienDien; }
+    public BigDecimal getTienDien() { return tienDien; }
+    public void setTienDien(BigDecimal tienDien) { this.tienDien = tienDien; }
 
     public Integer getSoNuocDung() { return soNuocDung; }
     public void setSoNuocDung(Integer soNuocDung) { this.soNuocDung = soNuocDung; }
 
-    public Double getTienNuoc() { return tienNuoc; }
-    public void setTienNuoc(Double tienNuoc) { this.tienNuoc = tienNuoc; }
+    public BigDecimal getTienNuoc() { return tienNuoc; }
+    public void setTienNuoc(BigDecimal tienNuoc) { this.tienNuoc = tienNuoc; }
 
-    public Double getTongTien() { return tongTien; }
-    public void setTongTien(Double tongTien) { this.tongTien = tongTien; }
+    public BigDecimal getTongTien() { return tongTien; }
+    public void setTongTien(BigDecimal tongTien) { this.tongTien = tongTien; }
 }
