+++ src/main/java/com/btl/server/repository/KhachHangRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface KhachHangRepository extends JpaRepository<KhachHang, Integer> {
+public interface KhachHangRepository extends JpaRepository<KhachHang, Long> {
    
     Optional<KhachHang> findByTaiKhoan(TaiKhoan taiKhoan);
     Optional<KhachHang> findBySoCccd(String soCccd);
